swiss tournament
This project is the second part of the Udacity Fullstack Nanodegree
program. It is a python module which keeps track of players in a
tournament. It test various functionality of the module, including
Table/View creation, insertion and deletion, and also tests a
Swiss-style tournament arrangement, which produces a definite winner in
a few rounds, no-one is eliminated and everyone wins at least one round.